+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
+### Added
+- Allow custom config human reference tracks, e.g. for local emergency copies when UCSC is unreachable (#6578)
 ### Changed
 - Changelog enforcement specific to the `unreleased` changelog section (#6216)